+++ snack-lib/default.nix
@@ -2,6 +2,8 @@
 # TODO: currently single out derivations prepend the PWD to the path
 # TODO: make sure that filters for "base" are airtight
 { pkgs
+, ghc-version ? "ghc822"
+, ghcWithPackages ? pkgs.haskell.packages.${ghc-version}.ghcWithPackages
 }:
 
 with pkgs;
@@ -16,7 +18,7 @@ with (callPackage ./package-spec.nix {});
 with (callPackage ./hpack.nix {});
 
 let
-  ghcWith = deps: haskellPackages.ghcWithPackages
+  ghcWith = deps: ghcWithPackages
     (ps: map (p: ps.${p}) deps);
 
   # Assumes the package description describes an executable
@@ -65,12 +67,12 @@ let
         exe_path = "${drv.out}/${drv.relExePath}";
       };
 
-  inferSnackBuild = snackNix: mkPackage (import snackNix);
+  inferSnackBuild = packageNix: mkPackage (import packageNix);
 
-  inferSnackGhci = snackNix: writeText "snack-ghci-json"
+  inferSnackGhci = packageNix: writeText "snack-ghci-json"
     ( builtins.toJSON (
     let
-      pkgSpec = mkPackageSpec (import snackNix);
+      pkgSpec = mkPackageSpec (import packageNix);
       drv =
         if builtins.isNull pkgSpec.packageMain
         then ghciWithModules ghcWith (libraryModSpecs pkgSpec)
@@ -102,7 +104,7 @@ let
     ( builtins.toJSON (
     let
       pkgSpecs = hpackSpecs packageYaml;
-      pkgSpec = mkPackageSpec (import snackNix);
+      pkgSpec = mkPackageSpec (import packageNix);
       drv =
         let exeSpecs = builtins.attrValues pkgSpecs.executables;
         in
@@ -120,7 +122,7 @@ let
       }
     ));
 
-  snackSpec = snackNix: mkPackageSpec (import snackNix);
+  snackSpec = packageNix: mkPackageSpec (import packageNix);
   hpackSpecs = packageYaml:
     let
       descrs = pkgDescrsFromHPack packageYaml;