+++ tests/extended-config/snack.nix
@@ -1,0 +1,4 @@
+rec {
+  pkgs = import ./nix;
+  ghcWithPackages = pkgs.haskellPackages.ghcWithPackages;
+}